Move some tests to external
